+++ backend/src/main/java/vn/edu/ptit/int1433/training/exception/SubmissionValidationException.java
@@ -1,0 +1,7 @@
+package vn.edu.ptit.int1433.training.exception;
+
+public class SubmissionValidationException extends RuntimeException {
+    public SubmissionValidationException(String message) {
+        super(message);
+    }
+}